+++ dbus_remote_proxy/lib/dbus_remote_proxy.dart
@@ -4,179 +4,13 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:dbus/dbus.dart';
+import 'package:dbus_remote_proxy/dbus_utils.dart';
+import 'package:dbus_remote_proxy/dbus_value_converter.dart';
 import 'package:flutter/foundation.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
 
-const String systemService = 'system';
-
-// convert native value to DBusValue
-// DBusValue fromNativeValue(dynamic value) {
-//   if (value is String) {
-//     return DBusString(value);
-//   } else if (value is int) {
-//     if (value.bitLength <= 8) return DBusByte(value);
-//     if (value.bitLength <= 16) return DBusInt16(value);
-//     if (value.bitLength <= 32) return DBusInt32(value);
-//     return DBusInt64(value);
-//   } else if (value is bool) {
-//     return DBusBoolean(value);
-//   } else if (value is double) {
-//     return DBusDouble(value);
-//   } else if (value is List) {
-//     if (value.isNotEmpty && value.first is Map) {
-//       return DBusArray(
-//         DBusSignature('a{ss}'),
-//         value.map((v) => fromNativeValue(v)).toList(),
-//       );
-//     } else if (value.isNotEmpty && value.first is List) {
-//       return DBusArray(
-//         DBusSignature('aa{ss}'),
-//         value.map((v) => fromNativeValue(v)).toList(),
-//       );
-//     } else {
-//       return DBusArray(
-//         DBusSignature(value.isNotEmpty
-//             ? fromNativeValue(value.first).signature.value
-//             : 'v'),
-//         value.map((v) => fromNativeValue(v)).toList(),
-//       );
-//     }
-//   } else if (value is Map) {
-//     return DBusDict(
-//       DBusSignature('s'),
-//       DBusSignature('s'),
-//       value.map((k, v) => MapEntry(
-//             DBusString(k.toString()),
-//             DBusString(v.toString()),
-//           )),
-//     );
-//   } else {
-//     throw Exception('Unsupported native value type: ${value.runtimeType}');
-//   }
-// }
-
-// Convert native value to DBusValue with support for variants and structs
-// TODO: should improve this method to cover all of replySignature
-DBusValue fromNativeValue(dynamic value, {DBusSignature? expectedSignature}) {
-  if (expectedSignature != null) {
-    if (expectedSignature.value.startsWith('a') && value is List) {
-      final childSignature =
-          DBusSignature(expectedSignature.value.substring(1));
-      return DBusArray(
-        childSignature,
-        value
-            .map((v) => fromNativeValue(v, expectedSignature: childSignature))
-            .toList(),
-      );
-    }
-    if (expectedSignature.value == '(oa{sv})' && value is List) {
-      if (value.length != 2) {
-        throw Exception('Struct (oa{sv}) need to have 2 elements');
-      }
-      return DBusStruct([
-        fromNativeValue(value[0], expectedSignature: DBusSignature('o')),
-        fromNativeValue(value[1], expectedSignature: DBusSignature('a{sv}')),
-      ]);
-    }
-
-    if (expectedSignature.value == 'a{sv}' && value is Map) {
-      return DBusDict(
-        DBusSignature('s'),
-        DBusSignature('v'),
-        value.map((k, v) => MapEntry(
-              DBusString(k.toString()),
-              fromNativeValue(v, expectedSignature: DBusSignature('v')),
-            )),
-      );
-    }
-
-    if (expectedSignature.value == 'v') {
-      if (value is String) {
-        return DBusVariant(DBusString(value));
-      } else if (value is int) {
-        if (value.bitLength <= 8) return DBusVariant(DBusByte(value));
-        if (value.bitLength <= 16) return DBusVariant(DBusInt16(value));
-        if (value.bitLength <= 32) return DBusVariant(DBusInt32(value));
-        return DBusVariant(DBusInt64(value));
-      } else if (value is bool) {
-        return DBusVariant(DBusBoolean(value));
-      } else if (value is double) {
-        return DBusVariant(DBusDouble(value));
-      } else if (value is List) {
-        if (value.isEmpty) return DBusVariant(DBusArray.string([]));
-        final childSignature = fromNativeValue(value.first).signature;
-        return DBusVariant(DBusArray(
-          childSignature,
-          value.map((v) => fromNativeValue(v)).toList(),
-        ));
-      } else if (value is Map) {
-        return DBusVariant(DBusDict(
-          DBusSignature('s'),
-          DBusSignature('v'),
-          value.map((k, v) => MapEntry(
-                DBusString(k.toString()),
-                fromNativeValue(v, expectedSignature: DBusSignature('v')),
-              )),
-        ));
-      } else {
-        throw Exception('Unsupported variant value type: ${value.runtimeType}');
-      }
-    }
-
-    if (expectedSignature.value == 'o' && value is String) {
-      return DBusObjectPath(value);
-    }
-  }
-
-  if (value is String) {
-    return DBusString(value);
-  } else if (value is int) {
-    if (value.bitLength <= 8) return DBusByte(value);
-    if (value.bitLength <= 16) return DBusInt16(value);
-    if (value.bitLength <= 32) return DBusInt32(value);
-    return DBusInt64(value);
-  } else if (value is bool) {
-    return DBusBoolean(value);
-  } else if (value is double) {
-    return DBusDouble(value);
-  } else if (value is List) {
-    if (value.isEmpty) return DBusArray.string([]);
-    final firstElement = fromNativeValue(value.first);
-    return DBusArray(
-      firstElement.signature,
-      value.map((v) => fromNativeValue(v)).toList(),
-    );
-  } else if (value is Map) {
-    return DBusDict(
-      DBusSignature('s'),
-      DBusSignature('s'),
-      value.map((k, v) => MapEntry(
-            DBusString(k.toString()),
-            DBusString(v.toString()),
-          )),
-    );
-  } else {
-    throw Exception('Unsupported native value type: ${value.runtimeType}');
-  }
-}
-
-class DBusRemote {
-  static DBusRemoteObject? getObjectInstance(
-      String? type, String? name, DBusObjectPath? path) {
-    if (kIsWeb) {
-      return null;
-    }
-
-    DBusClient dbusClient =
-        type == systemService ? DBusClient.system() : DBusClient.session();
-
-    return DBusRemoteObject(dbusClient, name: name!, path: path!);
-  }
-}
-
 // A Proxy class contain DBusRemoteObject to abstraction callMethod layer
-class DBusRemoteObjectProxy {
-  final DBusRemoteObject? _object;
+class DBusRemoteObjectProxy extends DBusRemoteObject {
   WebSocketChannel? _channel;
   final bool _useWebSocket;
   final StreamController<dynamic> _messageController =
@@ -193,8 +27,8 @@ class DBusRemoteObjectProxy {
 
   DBusRemoteObjectProxy(
       {required this.type, required this.name, required this.path})
-      : _object = DBusRemote.getObjectInstance(type, name, path),
-        _useWebSocket = kIsWeb {
+      : _useWebSocket = kIsWeb,
+        super(DBusUtils.getDBusClientByType(type), name: name, path: path) {
     pathValue = path.value;
     if (_useWebSocket) {
       _connectToWebSocket();
@@ -223,30 +57,31 @@ class DBusRemoteObjectProxy {
     }
   }
 
-  Future<DBusMethodResponse> callMethod(
-    String? interface,
-    String member,
-    Iterable<DBusValue> values, {
-    DBusSignature? replySignature,
-  }) async {
+  @override
+  Future<DBusMethodSuccessResponse> callMethod(
+      String? interface, String name, Iterable<DBusValue> values,
+      {DBusSignature? replySignature,
+      bool noReplyExpected = false,
+      bool noAutoStart = false,
+      bool allowInteractiveAuthorization = false}) async {
     if (_useWebSocket) {
       if (_channel == null) {
         throw Exception('WebSocket not connected');
       }
 
       final request = jsonEncode({
-        'serviceName': name,
+        'serviceName': this.name,
         'serviceType': type,
         'path': pathValue,
         'interface': interface,
-        'member': member,
+        'member': name,
         'values': values.map((v) => v.toNative()).toList(),
         'replySignature': replySignature?.value,
       });
       _channel!.sink.add(request);
 
       // wait message from StreamController
-      final completer = Completer<DBusMethodResponse>();
+      final completer = Completer<DBusMethodSuccessResponse>();
       late StreamSubscription subscription;
       subscription = _messageController.stream.listen(
         (event) {
@@ -255,8 +90,8 @@ class DBusRemoteObjectProxy {
             final result = jsonDecode(message);
             if (result['status'] == 'success') {
               final returnValues = (result['returnValues'] as List)
-                  .map((v) =>
-                      fromNativeValue(v, expectedSignature: replySignature))
+                  .map((v) => DBusValueConverter.fromNativeValue(v,
+                      expectedSignature: replySignature))
                   .toList();
               completer.complete(DBusMethodSuccessResponse(returnValues));
             } else {
@@ -287,11 +122,11 @@ class DBusRemoteObjectProxy {
         },
       );
     } else {
-      if (_object == null) {
-        throw Exception('D-Bus object not initialized');
-      }
-      return await _object.callMethod(interface, member, values,
-          replySignature: replySignature);
+      return await super.callMethod(interface, name, values,
+          replySignature: replySignature,
+          noReplyExpected: noReplyExpected,
+          noAutoStart: noAutoStart,
+          allowInteractiveAuthorization: allowInteractiveAuthorization);
     }
   }